refactor: support add from url
